New docs section for compressed
