+++ src-docs/src/routes.js
@@ -96,6 +96,8 @@ import { FormControlsExample } from './views/form_controls/form_controls_example
 
 import { FormLayoutsExample } from './views/form_layouts/form_layouts_example';
 
+import { FormCompressedExample } from './views/form_compressed/form_compressed_example';
+
 import { FormValidationExample } from './views/form_validation/form_validation_example';
 
 import { HeaderExample } from './views/header/header_example';
@@ -328,6 +330,7 @@ const navigation = [
     items: [
       FormControlsExample,
       FormLayoutsExample,
+      FormCompressedExample,
       FormValidationExample,
       SuperSelectExample,
       ComboBoxExample,
